requirements.txt
flask==3.1.3
numpy==1.24.3
pandas==2.3.3
scikit-learn==1.7.2
tensorflow-cpu==2.13.0
nltk==3.9.3
matplotlib==3.10.8
seaborn==0.13.2
pytesseract==0.3.13
Pillow==12.1.1
PyPDF2==3.0.1
python-docx==1.2.0